refactor: index folders for each entity (#751)
